Frontend add documentation and fix PR comments
