+++ frontend/src/stores/filterState/FilterStateStore.ts
@@ -1,28 +1,26 @@
+/* istanbul ignore file */
 import { injectable } from 'inversify';
-import SimpleStore from './SimpleStore';
+import SimpleStore from '../SimpleStore';
+import {
+  FilterLineState,
+  FilterPropertyState,
+  FilterState,
+} from './FilterState';
 
-export interface FilterPropertyState {
-  name: string;
-  values: string[];
-}
-
-export interface FilterLineState {
-  type: string;
-  isActive: boolean;
-  propertyFilters: FilterPropertyState[];
-}
-
-export interface FilterState {
-  edges: FilterLineState[];
-  nodes: FilterLineState[];
-}
-
+/**
+ * Contains the current state of the filter.
+ */
 @injectable()
 export default class FilterStateStore extends SimpleStore<FilterState> {
   protected getInitialValue(): FilterState {
     return { edges: [], nodes: [] };
   }
 
+  /**
+   * Toggles the {@link isActive} property of the specified filter line.
+   * @param filterLineType - filter line which {@link isActive} property is toggled
+   * @param entity - specifies whether this entity is a node or an edge
+   */
   public toggleFilterLineActive(
     filterLineType: string,
     entity: 'node' | 'edge'
@@ -39,6 +37,14 @@ export default class FilterStateStore extends SimpleStore<FilterState> {
     this.setState(this.getValue());
   }
 
+  /**
+   * Adds a new {@link FilterPropertyState} to the specified filter line. If there already
+   * exists a {@link FilterPropertyState} with the given name in the filter line, the
+   * {@link FilterPropertyState} is overwritten with the new one.
+   * @param stateToAdd - the {@link FilterPropertyState} to be added
+   * @param filterLineType - filter line which the {@link FilterPropertyState} is added to
+   * @param entity - specifies whether this entity is a node or an edge
+   */
   public addFilterPropertyState(
     stateToAdd: FilterPropertyState,
     filterLineType: string,
@@ -63,6 +69,13 @@ export default class FilterStateStore extends SimpleStore<FilterState> {
     this.setState(this.getValue());
   }
 
+  /**
+   * Gets the values of a {@link FilterPropertyState} in the specified filter line if they exist.
+   * Otherwise undefined is returned.
+   * @param filterLineType - filter line which is searched for
+   * @param filterPropertyName - name of the {@link FilterPropertyState}
+   * @param entity - specifies whether this entity is a node or an edge
+   */
   public getPropertyStateValues(
     filterLineType: string,
     filterPropertyName: string,
@@ -78,20 +91,17 @@ export default class FilterStateStore extends SimpleStore<FilterState> {
     filterLineType: string,
     entity: 'node' | 'edge'
   ): FilterLineState | undefined {
-    let result: FilterLineState | undefined;
-
     const filterState = this.getValue();
     const entityLineStates =
       entity === 'node' ? filterState.nodes : filterState.edges;
 
     for (const entityLine of entityLineStates) {
       if (entityLine.type === filterLineType) {
-        result = entityLine;
-        break;
+        return entityLine;
       }
     }
 
-    return result;
+    return undefined;
   }
 
   private searchForPropertyState(
@@ -99,8 +109,6 @@ export default class FilterStateStore extends SimpleStore<FilterState> {
     filterPropertyName: string,
     entity: 'node' | 'edge'
   ) {
-    let result: FilterPropertyState | undefined;
-
     const line: FilterLineState | undefined = this.searchForLineState(
       filterLineType,
       entity
@@ -109,10 +117,11 @@ export default class FilterStateStore extends SimpleStore<FilterState> {
     if (line) {
       for (const propertyFilters of line.propertyFilters) {
         if (propertyFilters.name === filterPropertyName) {
-          result = propertyFilters;
+          return propertyFilters;
         }
       }
     }
-    return result;
+
+    return undefined;
   }
 }
